# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Games Review Website Rest API DOCUMENTATION",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "arifhossain393@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/comment": {
            "post": {
                "tags": [
                    "User Comment"
                ],
                "summary": "Create comment",
                "description": "Create Comment",
                "operationId": "52b8b691fa7522bb5d041b02158037f4",
                "requestBody": {
                    "description": "Games objects",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "games_id",
                                    "r_star",
                                    "u_email"
                                ],
                                "properties": {
                                    "games_id": {
                                        "description": "Games Id",
                                        "type": "integer",
                                        "example": "Games Id"
                                    },
                                    "r_star": {
                                        "description": "Games star",
                                        "type": "integer",
                                        "example": "games-star"
                                    },
                                    "u_name": {
                                        "description": "User Name",
                                        "type": "string",
                                        "example": "Jhon Doe"
                                    },
                                    "u_email": {
                                        "description": "User Email",
                                        "type": "string",
                                        "example": "jhon@gmail.com"
                                    },
                                    "comment": {
                                        "description": "Comment",
                                        "type": "text",
                                        "example": "Write you comment"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/games-comments": {
            "get": {
                "tags": [
                    "Games Comment"
                ],
                "summary": "Get all games Comment for REST API",
                "description": "Multiple status values can be provided with comma separated string",
                "operationId": "c3bd83dcb80b7f0301843d1eac3c3c88",
                "parameters": [
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Per page count",
                        "required": false,
                        "explode": true,
                        "schema": {
                            "type": "integer",
                            "default": "10"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search by title",
                        "required": false,
                        "explode": true,
                        "schema": {
                            "type": "string",
                            "default": ""
                        }
                    },
                    {
                        "name": "orderBy",
                        "in": "query",
                        "description": "Order By column name",
                        "required": false,
                        "explode": true,
                        "schema": {
                            "type": "string",
                            "default": "id"
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Order ordering - asc or desc",
                        "required": false,
                        "explode": true,
                        "schema": {
                            "type": "string",
                            "default": "desc"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/games-list": {
            "get": {
                "tags": [
                    "Games Lists"
                ],
                "summary": "Get all games for REST API",
                "description": "Multiple status values can be provided with comma separated string",
                "operationId": "b7537694432340a50074088a197fdd2b",
                "parameters": [
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Per page count",
                        "required": false,
                        "explode": true,
                        "schema": {
                            "type": "integer",
                            "default": "10"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search by title",
                        "required": false,
                        "explode": true,
                        "schema": {
                            "type": "string",
                            "default": ""
                        }
                    },
                    {
                        "name": "orderBy",
                        "in": "query",
                        "description": "Order By column name",
                        "required": false,
                        "explode": true,
                        "schema": {
                            "type": "string",
                            "default": "id"
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Order ordering - asc or desc",
                        "required": false,
                        "explode": true,
                        "schema": {
                            "type": "string",
                            "default": "desc"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Games Lists"
                ],
                "summary": "Create Games",
                "description": "Create games",
                "operationId": "57a80caded2aaed7cc8624de8ab564ce",
                "requestBody": {
                    "description": "Games objects",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "price"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "Games title",
                                        "type": "string",
                                        "example": "Games title"
                                    },
                                    "slug": {
                                        "description": "Games slug",
                                        "type": "string",
                                        "example": "games-title"
                                    },
                                    "price": {
                                        "description": "games price",
                                        "type": "integer",
                                        "example": "200"
                                    },
                                    "image": {
                                        "description": "games image",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "description": {
                                        "description": "Description",
                                        "type": "text",
                                        "example": "description"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/games-list/{id}": {
            "get": {
                "tags": [
                    "Games Lists"
                ],
                "summary": "Get games detail",
                "description": "Get games detail",
                "operationId": "9d09ef6818b5dac7d0b54fd655743ca6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "games-list id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Games Lists"
                ],
                "summary": "Update games",
                "description": "Update games",
                "operationId": "f60dbcef11760bbd3b9836c6675468d2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "games-list id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "_method",
                        "in": "query",
                        "description": "request method",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "PUT"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Games objects",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "price",
                                    "slug"
                                ],
                                "properties": {
                                    "id": {
                                        "description": "games-list id",
                                        "type": "integer",
                                        "example": "games-list id"
                                    },
                                    "title": {
                                        "description": "games title",
                                        "type": "string",
                                        "example": "games title"
                                    },
                                    "slug": {
                                        "description": "games slug",
                                        "type": "string",
                                        "example": "games-title"
                                    },
                                    "price": {
                                        "description": "games price",
                                        "type": "integer",
                                        "example": "200"
                                    },
                                    "image": {
                                        "description": "games image",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "description": {
                                        "description": "Description",
                                        "type": "text",
                                        "example": "description"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Games Lists"
                ],
                "summary": "Delete games",
                "description": "Delete games",
                "operationId": "5f1a4ece4ef956794770bf5d6a02e6d5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "games id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "games not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login",
                "description": "Login to system.",
                "operationId": "login",
                "requestBody": {
                    "description": "Pet object that needs to be added to the store",
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "User Email",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "User password",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register",
                "description": "Register to system.",
                "operationId": "register",
                "requestBody": {
                    "description": "Pet object that needs to be added to the store",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "User Name",
                                        "type": "string",
                                        "example": "Jhon Doe"
                                    },
                                    "email": {
                                        "description": "User Email",
                                        "type": "string",
                                        "example": "jhon@example.com"
                                    },
                                    "password": {
                                        "description": "User password",
                                        "type": "string",
                                        "example": "12345678"
                                    },
                                    "password_confirmation": {
                                        "description": "User confirm password",
                                        "type": "string",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        }
    }
}